feat: broad hardening pass — graph search, telemetry pipeline, cache/WAL resilience, and conformance fixes
A wide-reaching quality and robustness pass touching 31 files across
the frankensearch workspace (+742 lines, -141 lines):

New modules:
- frankensearch-core/src/graph.rs: new graph-based search support
  module for knowledge-graph-augmented retrieval
- frankensearch-index/src/repro_bug.rs: reproduction harness for an
  index corruption edge case under concurrent WAL compaction

Search engine core:
- searcher.rs: major expansion (+176 lines) — add lifecycle/resource
  telemetry hooks, federated sub-searcher coordination improvements,
  and graph-augmented query path
- search.rs: expand index search with better scoring normalization
  and edge-case handling for empty segments (+90 lines)
- warmup.rs: add segment pre-warming with memory-budget awareness
  (+62 lines)
- wal.rs: harden write-ahead log with improved fsync discipline,
  compaction checkpoint recovery, and corrupted-entry skip (+80 lines)

Telemetry & observability:
- telemetry_adapter_common.sh: add 64-line shared telemetry shell
  library for E2E test scripts (structured JSONL emission, stage
  accounting, timeout handling)
- metrics.rs: expand metric counters with graph-search and cache
  eviction tracking
- pressure.rs / pressure_sensing.rs: improve memory-pressure
  detection with smoothed readings and configurable thresholds

Cache & configuration:
- cache.rs: add TTL-aware eviction and cache-hit-ratio telemetry
- config.rs: extend FSFS config with pressure-sensing tunables
- types.rs: add graph-search result types to the core type system

Fusion & reranking:
- conformal.rs: expand conformal prediction calibration with
  confidence-interval telemetry
- model2vec_embedder.rs: refactor embedding batch pipeline for
  better error propagation
- pipeline.rs: add reranking telemetry emission
- rrf.rs, federated.rs, lexical/lib.rs: minor import/use fixes

Tests & benchmarks:
- interaction_integration.rs, interaction_unit.rs: fix imports
- cross_component.rs, bench_quick.rs, validate_fusion.rs,
  search_bench.rs: fix imports for new types
- optimize_params/main.rs: align with updated config API

Beads: sync issue tracker state

+++ crates/frankensearch-tui/src/interaction.rs
@@ -203,9 +203,14 @@ impl InteractionLatencyHooks {
         }
     }
 
+    /// Returns the total component budget in milliseconds.
+    ///
+    /// Uses `u32` to avoid overflow when summing three `u16` fields.
     #[must_use]
-    pub const fn component_budget_ms(self) -> u16 {
-        self.input_to_route_ms + self.route_to_state_ms + self.state_to_render_ms
+    pub const fn component_budget_ms(self) -> u32 {
+        self.input_to_route_ms as u32
+            + self.route_to_state_ms as u32
+            + self.state_to_render_ms as u32
     }
 
     fn validate(self, surface: InteractionSurfaceKind) -> Result<(), ShowcaseInteractionSpecError> {
@@ -219,7 +224,7 @@ impl InteractionLatencyHooks {
                 "latency hooks must all be > 0".to_owned(),
             ));
         }
-        if self.component_budget_ms() > self.frame_budget_ms {
+        if self.component_budget_ms() > u32::from(self.frame_budget_ms) {
             return Err(ShowcaseInteractionSpecError::InvalidLatencyBudget(
                 surface,
                 format!(
@@ -743,7 +748,7 @@ mod tests {
         for surface in &spec.surfaces {
             assert!(
                 surface.latency_hooks.component_budget_ms()
-                    <= surface.latency_hooks.frame_budget_ms
+                    <= u32::from(surface.latency_hooks.frame_budget_ms)
             );
         }
     }
